working on the like button
