+++ src/index.css
@@ -12,5 +12,10 @@ code {
     monospace;
 }
 
+.text-area {
+  min-width: 500px;
+  min-height: 100px;
+}
+
 .post-form {
 }
